+++ PlantVsZombies/Game/Zombie/ConeZombie.cpp
@@ -22,7 +22,7 @@ void ConeZombie::HelmDrop()
 void ConeZombie::CheckHelmImage()
 {
 	if (mHelmType == HelmType::HELMTYPE_NONE) return;
-	if (mHelmStage == ArmorBrokenState::NO_BROKEN && mHelmHealth <= mHelmMaxHealth * 2 / 3) {
+	if (mHelmStage == ArmorBrokenState::NO_BROKEN && mHelmHealth <= static_cast<long long>(mHelmMaxHealth) * 2 / 3) {
 		mHelmStage = ArmorBrokenState::A_LITTLE_BROKEN;
 		mAnimator->SetTrackImage("anim_cone", ResourceManager::GetInstance().
 			GetTexture("IMAGE_ZOMBIE_CONE2"));
